requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
openai==1.51.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pydantic==2.9.2
numpy==1.26.4